Move implementation of Instantiate to decls.rs
