+++ lang/syntax/src/trees/ust/mod.rs
@@ -1,5 +1,4 @@
 mod def;
 mod subst;
-pub mod util;
 
 pub use def::*;
